pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taskcask"
version = "0.0.1"
authors = [
  { name="Ignytis", email="155588001+ignytis@users.noreply.github.com" },
]
description = "A process launcher tool"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "pydantic>=2.10.5,<3.0.0",
]

[project.optional-dependencies]
cli = [
  "click>=8.1.8,<9.0.0",
]
yaml = [
  "pyyaml>=6.0.0,<7.0.0",
]

[project.scripts]
tcask = "taskcask.cli.main:cmd_group_main"

[project.entry-points."taskcask.task_templates.class_loaders"]
stdlib = "taskcask.stdlib.task_templates.class_loader:ClassLoader"

[project.entry-points."taskcask.task_templates.definition_loaders"]
stdlib = "taskcask.stdlib.task_templates.definition_loaders.dir_yaml_loader:Loader"

[project.entry-points."taskcask.executors.class_loaders"]
stdlib = "taskcask.stdlib.executors.class_loader:ExecutorClassLoader"

[project.urls]
Homepage = "https://github.com/ignytis/taskcask"

[tool.hatch.build]
only-packages = true
